#3280 TableLocator ready and new models being called from Carto::xxx controllers
