+++ app/controllers/api/json/geocodings_controller.rb
@@ -154,6 +154,6 @@ class Api::Json::GeocodingsController < Api::ApplicationController
   protected
 
   def load_table
-    @table = ::Table.get_by_id_or_name(params.fetch('table_name'), current_user)
+    @table = Helpers::TableLocator.get_by_id_or_name(params.fetch('table_name'), current_user)
   end
 end
