+++ app/controllers/api/json/tables_controller.rb
@@ -110,7 +110,7 @@ class Api::Json::TablesController < Api::ApplicationController
   protected
 
   def load_table
-    @table = ::Table.get_by_id_or_name(params.fetch('id'), current_user)
+    @table = Helpers::TableLocator.get_by_id_or_name(params.fetch('id'), current_user)
   end
 end
