+++ app/controllers/carto/api/records_controller.rb
@@ -26,7 +26,7 @@ module Carto
 
       def table
         if @table.nil?
-          @table = ::Table.get_by_id_or_name(params[:table_id], current_user)  
+          @table = Carto::Helpers::TableLocator.get_by_id_or_name(params[:table_id], current_user)  
           raise RecordNotFound if @table.nil?
         end
         @table
